Fixes advanceLine not moving down enough.
The font height measures from the baseline to the highest point in the
font. The font margin is supposed to add space between the two lines. The font descent (added in
this commit) measures from the baseline to the lowest point in the font.

See https://github.com/labouseur/TSOS-2014/issues/6 for a fuller explaination.